Fix issue with existing ld

+++ build.sh
@@ -23,7 +23,9 @@ RELEASE="$(rpm -E %fedora)"
 # systemctl enable podman.socket
 
 # Install tmff2 driver.
-ln -s /usr/bin/ld.bfd /usr/bin/ld
+if [[ ! -f /usr/bin/ld ]]; then
+  ln -s /usr/bin/ld.bfd /usr/bin/ld
+fi
 mkdir -p /tmp/hid-tmff2
 cd /tmp/hid-tmff2
 git clone --recurse-submodules https://github.com/hellfur/hid-tmff2.git
